Model deviance

Model deviance

# S3 method for Owl
deviance(object, ...)

Arguments

object
an object of class 'Owl'.

...
ignored

Value

For Gaussian models this is twice the residual sums of squares. For all other models, two times the negative loglikelihood is returned.

Examples

fit <- owl(heart$x, heart$y, family = "binomial")
deviance(fit)

#>  [1] 370.9593 352.9908 333.8938 317.7167 303.8756 291.8126 281.0661 270.5324
#>  [9] 260.9091 252.4471 244.9932 238.4299 232.6748 227.6071 222.8242 217.4378
#> [17] 212.5163 207.8711 203.7203 199.6878 196.2031 192.9188 189.9912 187.4567
#> [25] 185.2710 183.2485 181.5044 179.9265 178.5430 177.3053 176.1985 175.2377
#> [33] 174.4021 173.6924 173.0818 172.5474 172.0739 171.6674 171.2815 170.9316
#> [41] 170.6394 170.3815 170.1597 169.9700 169.8087 169.6730 169.5569 169.4587
#> [49] 169.3712 169.2973 169.2350 169.1812 169.1352 169.0959 169.0633 169.0359
#> [57] 169.0128 168.9936 168.9775 168.9640 168.9527 168.9433 168.9354 168.9289
#> [65] 168.9234 168.9189 168.9150 168.9119 168.9092 168.9070 168.9052 168.9037