Generalized Linear Models Penalized with the SLOPE (OWL) Norm

SLOPE (Sorted L-One Penalized Estimation) is an extension of the Lasso. Unlike the latter, however, SLOPE uses a non-increasing sequence of \(\lambda\)---one for each coefficient. The penalty term looks like

owl(
  x,
  y,
  family = c("gaussian", "binomial", "multinomial", "poisson"),
  intercept = TRUE,
  standardize_features = TRUE,
  sigma = NULL,
  lambda = c("gaussian", "bhq", "oscar"),
  lambda_min_ratio = if (n < p) 0.01 else 1e-04,
  n_sigma = 100,
  q = 0.1 * min(1, n/p),
  screening = TRUE,
  tol_dev_change = 1e-05,
  tol_dev_ratio = 0.995,
  max_variables = n * m,
  max_passes = 1e+06,
  tol_rel_gap = 1e-05,
  tol_infeas = 1e-04,
  diagnostics = FALSE,
  verbosity = 0
)

Arguments

x
the feature matrix, which can be either a dense matrix of the standard matrix class, or a sparse matrix inheriting from Matrix::sparseMatrix Data frames will be converted to matrices internally.

y
the response. For Gaussian models this must be numeric; for binomial models, it can be a factor.

family
response type. See Families for details.

intercept
whether to fit an intercept

standardize_features
whether to standardize features (predictors)

sigma
scale of lambda sequence

lambda
either a character vector indicating the method used to construct the lambda path or a vector with length equal to the number of coefficients in the model

lambda_min_ratio
smallest value for lambda as a fraction of lambda_max

n_sigma
length of regularization path

q
shape of lambda sequence

screening
whether the strong rule for SLOPE be used to screen variables for inclusion

tol_dev_change
the regularization path is stopped if the fractional change in deviance falls below this value. Note that this is automatically set to 0 if a sigma is manually entered

tol_dev_ratio
the regularization path is stopped if the deviance ratio \(1 - \mathrm{deviance}/\mathrm{(null-deviance)}\) is above this threshold

max_variables
criterion for stoping the path in terms of the maximum number of unique, nonzero coefficients in absolute value in model

max_passes
maximum number of passes for optimizer

tol_rel_gap
stopping criterion for the duality gap

tol_infeas
stopping criterion for the level of infeasibility

diagnostics
should diagnostics be saved for the model fit (timings, primal and dual objectives, and infeasibility)

verbosity
level of verbosity for displaying output from the program. Setting this to 1 displays information on the path level, while setting it to 2 displays information also from inside the solver.

Value

An object of class "Owl" with the following slots:

coefficients

a three-dimensional array of the coefficients from the model fit, including the intercept if it was fit. There is one row for each coefficient, one column for each target (dependent variable), and one slice for each penalty.

nonzeros

a three-dimensional boolean array indicating whether a coefficient was zero or not

lambda

the lambda vector that when multiplied by a value in sigma gives the penalty vector at that point along the regularization path

sigma

the vector of sigma, indicating the scale of the lambda vector

class_names

a character vector giving the names of the classes for binomial and multinomial families

passes

the number of passes the solver took at each path

violations

the number of violations of the screening rule

active_sets

a list where each element indicates the indices of the coefficients that were active at that point in the regularization path

diagnostics

a data.frame of objective values for the primal and dual problems, as well as a measure of the infeasibility, time, and iteration. Only available if diagnostics = TRUE in the call to owl().

call

the call used for fitting the model

Details

$$ \sigma \sum_{i=j}^p \lambda_j |\beta|_{(j)} $$

The objective for each model is simply the loss function for each family plus a penalty term.

Families

Gaussian

The Gaussian model (Ordinary Least Squares) minimizes the following objective.

$$ ||y - X\beta||_2^2 $$

Binomial

The binomial model (logistic regression) has the following objective.

$$ \sum_{i=1}^n \log\left(1+ \exp\left(- y_i \left(x_i^T\beta + \alpha \right) \right) \right) $$

Poisson

In poisson regression, we use the following objective.

$$ -\sum_{i=1}^n \left(y_i\left(x_i^T\beta + \alpha\right) - \exp\left(x_i^T\beta + \alpha\right)\right) $$

Multinomial

In multinomial regression, we use the following objective.

$$ -\sum_{i=1}^n\left( \sum_{k=1}^m y_{ik}(x_i^T\beta_k + \alpha_k) - \log\sum_{k=1}^m \exp(x_i^T\beta_k + \alpha_k) \right) $$

See also

Examples

# Gaussian response

fit <- owl(bodyfat$x, bodyfat$y)

# Binomial response

fit <- owl(heart$x, heart$y, family = "binomial")

# Poisson response

fit <- owl(abalone$x, abalone$y, family = "poisson")

# Multinomial response

fit <- owl(wine$x, wine$y, family = "multinomial")