Return the loss of a owl model

This function is a unified interface to return various types of loss for a model fit with owl().

score(object, x, y, measure)

# S3 method for OwlGaussian
score(object, x, y, measure = c("deviance", "mse", "mae"))

# S3 method for OwlBinomial
score(object, x, y, measure = c("deviance", "mse", "mae", "accuracy", "auc"))

Arguments

object
an object of class "Owl"

x
feature matrix

y
response

measure
type of target measure. The default, "deviance", returns mean squared error for Gaussian models and deviance for Binomial models. "mse" returns mean squared error. "mae" returns mean absolute error, "accuracy" returns classification rate accuracy, and "auc" returns area under the ROC curve.

Value

The measure along the regularization path depending on the value in measure.

Examples

x <- subset(infert, select = c("induced", "age", "pooled.stratum"))
y <- infert$case

fit <- owl(x, y, family = "binomial")
score(fit, x, y, measure = "auc")

#>        p1        p2        p3        p4        p5        p6        p7        p8 
#> 0.5150055 0.4921504 0.4952903 0.5150055 0.5142753 0.5269076 0.5020811 0.5042716 
#>        p9       p10       p11       p12       p13       p14       p15       p16 
#> 0.5048558 0.4689303 0.5020811 0.5174151 0.5003286 0.4996714 0.4996714 0.5024461 
#>       p17       p18       p19       p20       p21       p22       p23       p24 
#> 0.5047097 0.5069003 0.5069003 0.5090179 0.5070464 0.5084337 0.5092369 0.5098211 
#>       p25       p26       p27       p28       p29       p30       p31       p32 
#> 0.5093830 0.5108434 0.5100402 0.5109164 0.5103322 0.5106243 0.5122307 0.5107704 
#>       p33       p34       p35       p36       p37       p38       p39       p40 
#> 0.5120847 0.5117196 0.5131800 0.5115005 0.5124498 0.5123038 0.5117196 0.5127419 
#>       p41       p42       p43       p44       p45       p46       p47       p48 
#> 0.5122307 0.5127419 0.5125958 0.5128149 0.5124498 0.5126689 0.5123038 0.5125958 
#>       p49       p50       p51       p52       p53       p54       p55       p56 
#> 0.5109894 0.5130340 0.5123768 0.5122307 0.5127419 0.5125228 0.5136181 0.5127419 
#>       p57       p58       p59       p60       p61       p62       p63       p64 
#> 0.5117926 0.5116466 0.5120847 0.5128149 0.5121577 0.5135451 0.5126689 0.5127419 
#>       p65       p66       p67       p68       p69       p70       p71       p72 
#> 0.5127419 0.5120117 0.5127419 0.5131800 0.5125958 0.5128149 0.5126689 0.5134721 
#>       p73       p74       p75       p76       p77       p78       p79       p80 
#> 0.5124498 0.5117196 0.5129609 0.5115736 0.5120117 0.5125228 0.5130340 0.5140562 
#>       p81       p82       p83       p84       p85       p86       p87       p88 
#> 0.5124498 0.5138372 0.5117926 0.5132530 0.5135451 0.5124498 0.5121577 0.5123768 
#>       p89       p90       p91       p92       p93       p94       p95       p96 
#> 0.5128149 0.5125958 0.5131800 0.5114275 0.5128149 0.5133991 0.5120847 0.5126689 
#>       p97       p98       p99      p100 
#> 0.5131800 0.5123768 0.5126689 0.5125958